requirements.txt
moviepy
python-dotenv
openai
faster_whisper
fastapi